nf-float v0.4.2 (#82)
The new version contains a bugfix that disable rerun on checkpoint
restore failure at the MMC side.  We will use the Nextflow's re-run
mechanism when error happens.  Because Nextflow will create a new
work directory for the re-run task and won't have any dirty files
of the previous failed run.